Module bamboo_core::entry

Re-exports

pub use decode::decode;

pub use verify::verify;

pub use verify_batch::verify_batch;

pub use publish::publish;

pub use crate::error::*;

Modules

decode

encode

publish

verify

verify_batch

Structs

Entry

Constants

MAX_ENTRY_SIZE
This is useful if you need to know at compile time how big an entry can get.

MAX_ENTRY_SIZE_

Functions

into_owned

is_lipmaa_required